+++ skills/vcpdeck/vcpdeck.cjs
@@ -40,7 +40,7 @@ var require_version = __commonJS({
     "use strict";
     Object.defineProperty(exports2, "__esModule", { value: true });
     exports2.VERSION = void 0;
-    exports2.VERSION = "0.6.4";
+    exports2.VERSION = "0.6.5";
   }
 });